requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
gpt4all==2.8.2
python-dotenv==1.0.1 